authentication successfully

+++ backend/controller/todo.controller.js
@@ -1,4 +1,5 @@
 const Todo = require("../models/todo");
+const User = require("../models/user");
 const fs = require("fs");
 const path = require("path");
 
@@ -6,7 +7,7 @@ const clearImage = require("../util/remove-image");
 
 exports.getTodos = async (req, res, next) => {
   try {
-    const todos = await Todo.find();
+    const todos = await Todo.find().populate("creator");
     if (todos.length == 0) {
       return res.status(404).json({ message: "No todo has been recorded" });
     }
@@ -19,13 +20,12 @@ exports.getTodos = async (req, res, next) => {
   }
 };
 
-// Get a single todo by ID
 exports.getTodoById = async (req, res, next) => {
   try {
     const todoId = req.params.id;
     const todo = await Todo.findById(todoId);
     if (!todo) {
-      return res.status(404).json({ message: "No todo has been recorded" });
+      return res.status(404).json({ message: "Could not find todo." });
     }
     return res.status(200).json({ message: "Fetching todo by ID", todo: todo });
   } catch (err) {
@@ -44,24 +44,32 @@ exports.addTodo = async (req, res, next) => {
   }
 
   const imageUrl = req.file.path;
-
-  const { title, errorDescription, fixCode, fixExplanation, creator } =
-    req.body;
+  const { title, errorDescription, fixCode, fixExplanation } = req.body;
 
   try {
+    const user = await User.findById(req.userId);
+    if (!user) {
+      return res.status(404).json({ message: "User not found!" });
+    }
+
     const todo = new Todo({
       title,
       imageUrl,
       errorDescription,
       fixCode,
       fixExplanation,
-      creator,
+      creator: req.userId,
     });
-    const result = await todo.save();
 
-    return res
-      .status(201)
-      .json({ message: "Add Todo successfully", todo: result });
+    await todo.save();
+
+    user.todos.push(todo);
+    await user.save();
+
+    return res.status(201).json({
+      message: "Add Todo successfully",
+      todo: todo,
+    });
   } catch (error) {
     if (!error.statusCode) {
       error.statusCode = 500;
@@ -72,8 +80,7 @@ exports.addTodo = async (req, res, next) => {
 
 exports.updateTodo = async (req, res, next) => {
   const id = req.params.id;
-  const { title, errorDescription, fixCode, fixExplanation, creator } =
-    req.body;
+  const { title, errorDescription, fixCode, fixExplanation } = req.body;
   let imageUrl = req.body;
 
   try {
@@ -107,12 +114,18 @@ exports.updateTodo = async (req, res, next) => {
       throw error;
     }
 
+    if (todo.creator.toString() !== req.userId) {
+      const error = new Error("Unauthorized! Please log in.");
+      error.status = 403;
+      throw error;
+    }
+
     todo.title = title;
     todo.imageUrl = imageUrl;
     todo.errorDescription = errorDescription;
     todo.fixCode = fixCode;
     todo.fixExplanation = fixExplanation;
-    todo.creator = creator;
+    todo.creator = req.userId;
     const result = await todo.save();
 
     return res
@@ -137,6 +150,12 @@ exports.deleteTodo = async (req, res, next) => {
     }
 
     clearImage(todo.imageUrl);
+
+    if (todo.creator.toString() !== req.userId) {
+      const error = new Error("Unauthorized! Please log in.");
+      error.status = 403;
+      throw error;
+    }
 
     await Todo.findByIdAndDelete(id);
     return res.status(200).json({ message: "Delete Todo successfully" });
